+++ Credixware/Internal/Hacks/Bunnyhop.h
@@ -13,12 +13,19 @@ namespace Bunnyhop {
 	}
 
 	void Tick(CUserCmd* pCmd) {
+		if (!Settings::Misc::bBunnyhop || !GetAsyncKeyState(Settings::Misc::BunnyhopKey)) {
+			return;
+		}
+
 		if (!LocalEntity || LocalEntity == NULL) {
 			Init();
 			return;
 		}
-		//int flags = *(int*)(*LocalEntity + 0x104);
-		//printf("Flags: %i\n", flags);
+
+		int fFlags = *reinterpret_cast<int*>((DWORD)LocalEntity->GetBaseEntity() + 0x104);
+		if (fFlags & FL_ONGROUND) {
+			pCmd->buttons |= IN_JUMP;
+		}
 	}
 }