+++ Credixware/Internal/Hacks/Hacks.h
@@ -5,6 +5,7 @@
 #include "Bunnyhop.h"
 #include "Chams.h"
 #include "ESP.h"
+#include "Events.h"
 #include "Menu.h"
 #include "Nightmode.h"
 #include "SilentAim.h"