+++ Credixware/Internal/Hooks/OverrideView.h
@@ -11,9 +11,8 @@ unsigned int* oOverrideViewFunc;
 
 void __fastcall hkOverrideView(void* ecx, void* edx, CViewSetup* setup) {
 	Thirdperson::Update(setup);
-	// FOV Changer
-	// NOTE: Really buggy when scoped in
-	//setup->fov = 100.0f;
+	// Uncomment for FOV Changer (NOTE: Buggy when scoped in)
+	//setup->fov = 120.0f;
 	oOverrideView(ecx, setup);
 }