+++ Credixware/Internal/SDK/C_BaseEntity.h
@@ -3,6 +3,8 @@
 #ifndef C_BASEENTITY_H
 #define C_BASEENTITY_H
 
+#include "IClientEntity.h"
+
 struct FireBulletsInfo_t;
 struct trace_t;
 class CTakeDamageInfo;
@@ -10,9 +12,13 @@ class ITraceFilter;
 class CDmgAccumulator;
 class CEntityMapData;
 
-/*class C_BaseEntity : public IClientEntity
+class C_BaseEntity : public IClientEntity
 {
-	
-};*/
+public:
+	void					SetModelIndex(int index) {
+		typedef void(__thiscall* Fn)(void*, int);
+		Utils::GetVFunc<Fn>(this, 75)(this, index);
+	}
+};
 
 #endif