+++ Credixware/Internal/main.cpp
@@ -31,3 +31,44 @@ BOOL WINAPI DllMain(HMODULE hModule, DWORD dwReason, LPVOID lpReserved) {
 	}
 	return TRUE;
 }
+
+/*
+if (true)
+    {
+        if (!I::EntList->GetClientEntityFromHandle(hWearables[0]))
+        {
+            static ClientClass* pClass;
+ 
+            if (!pClass)
+                pClass = I::BaseClient->GetAllClasses();
+                while (pClass)
+                {
+                    if (pClass->m_ClassID == CEconWearable)
+                        break;
+                    pClass = pClass->m_pNext;
+                }
+ 
+            int iEntry, iSerial;
+ 
+            pClass->m_pCreateFn(iEntry = (I::EntList->GetHighestEntityIndex() + 1), iSerial = (RandomInt(0x0, 0xFFF)));
+            hWearables[0] = iEntry | (iSerial << 16);
+ 
+            IClientEntity* pEnt = I::EntList->GetClientEntityFromHandle(hWearables[0]);
+ 
+            if (pEnt)
+            {
+                *MakePtr(int, pEnt, gOffset.m_iItemDefinitionIndex) = 5033;
+                *MakePtr(int, pEnt, gOffset.m_iItemIDHigh) = -1;
+                *MakePtr(int, pEnt, gOffset.m_iEntityQuality) = 4;
+                *MakePtr(int, pEnt, gOffset.m_iAccountID) = info.xuidlow;
+                *MakePtr(float, pEnt, gOffset.m_flFallbackWear) = 0.00000001f;
+                *MakePtr(int, pEnt, gOffset.m_nFallbackSeed) = 0;
+                *MakePtr(int, pEnt, gOffset.m_nFallbackStatTrak) = -1;
+                *MakePtr(int, pEnt, gOffset.m_nFallbackPaintKit) = 10026;
+                pEnt->SetModelIndex(I::ModelInfo->GetModelIndex("models/weapons/v_models/arms/glove_motorcycle/v_glove_motorcycle.mdl"));
+                pEnt->GetClientNetworkable()->PreDataUpdate(DATA_UPDATE_CREATED);
+            }
+        }
+        
+    }
+	*/